Overivew: Outlined the skeleton for the Program.cs class.
Program.cs is the test class for this program; this commit pushes changes outlining this fact.

Constants were created for all strings that relate to a special command.
The main loop was broken up into two parts: 1) Initialize and 2) Run.
IsSpecial() was created in order to determine if a string should be triggering a command (true) and thus run that command inside this method, or, if it should instead be added to the DoublyLinkedList (false).